feat(oklch): added oklch to color picker options
